fix Scoop-Core tests for Win10 (#hack)
* fixes tests by coercing the `test-unzip` return value back into a valid path string
* fixes #476

.# Discussion

For Win10 machines, `test-unzip` returns a bugged value, causing test failures.

Investigation shows that the combination of Pester, COM (used within unzip_old), and
Win10 leads to the odd/bugged return value. The return value contains the correct
string, but it is prefaced with a leading space which arises de-novo during the
execution return process from `test-unzip`. The value is correct and usable at the end
of `test-unzip` but is bugged upon receiving the value at the caller. Removing the COM
call from within `unzip_old` or using a Windows version earlier than Win10 fixes the
problem, but are obviously not useful as solutions.

Given the nature of the returned value and COM involvement, the real problem is likely
stack corruption. But it's not clear whether the issue is with Pester, Win10/PowerShell,
or possibly even some subtle bug within unzip_old.

This, admittedly hacky, coercion fixes the problem for all Windows versions and also
keeps Appveyor satisfied.